testing/web-platform/tests/workers/interfaces/WorkerUtils/importScripts/002.html














<!doctype html>
<title>importScripts resolving urls</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_false(e.data[0], 'first argument to importScripts ran');
    assert_true(e.data[1], 'space in URL in second argument to importScripts did not throw');
    assert_equals(e.data[2], 12, 'exception code');
    this.done();
  });
});
</script>
